Add -1 (raw mode) to just print raw prefix-list.
At present, the only ways to get a prefix list without any formatting
are to either use JSON output and pipe through "jq" or similar, or to
generate a normal Cisco IOS prefix list and "cut" the fifth field, or to
use a user-defined format of "-F '%n/%l\n'" which isn't so intuitive.

In order to make the lives of those who just want prefix lists easier,
I've added the "-1" option, which prints one prefix per line. I've
called it "raw" mode, but I imagine there is probably a better name for
it that I'll be very happy for it to change to!